auth/requirements.txt
flask==3.1.*
pyyaml==6.*
bcrypt==4.*
gunicorn==23.*
ldap3==2.*
